#476: Move dimming view from GuideScene to CommonViews
